drgn.helpers.linux.slab: retry on on freelist and partial slab list races
On busy slab caches, it's not hard to race with an update while walking
a freelist or partial slab list. We work around this in the crash
kmem -s command by adding retries, but the code generated by
kmem -s --drgn still hits this in our tests. Move the retries to the
slab helpers instead.

Signed-off-by: Omar Sandoval <osandov@osandov.com>