drgn.helpers.experimental.kmodify: fix/disallow write_object() for bit fields
write_object() is extremely broken for bit fields. It doesn't take the
bit offset or bit size into account, which means that it writes to the
wrong bits and zeroes additional bits, too.

Byte-aligned bit fields are easy once we consider the offset and size.
We need to be careful for the general case, as a non-atomic
read-modify-write is a huge footgun. Boris Burkov has a use case for
writing to a bit field of size 1, which we can easily do atomically on
x86-64, so implement that. We can implement other cases when they arise.

Fixes: 1652183026a8 ("Add experimental helpers for calling kernel functions and writing to memory")
Signed-off-by: Omar Sandoval <osandov@osandov.com>