drgn.helpers.experimental.kmodify: add {set,clear}_bit()
We added the machinery needed for setting/clearing bits atomically for
bit fields, so expose it for bitmaps, too. Big-endian bit numbering is
annoying, but we don't need to worry about it for now since we only
support kmodify on x86-64.

These strictly take an unsigned long * to avoid future confusion with
big-endian bit numbering.

Signed-off-by: Omar Sandoval <osandov@osandov.com>