crash: port rd command
The rd command allows reading and formatting kernel memory in a variety
of ways. In addition to being its own command, it also provides
functionality that other commands (like bt) will use, so it's important
to provide an API for its functionality as well. But since it is just
implementing crash-compatible behavior, it's not exposed as a helper.

Thanks to the recent refactor to identify_address_all(), it is possible
to implement the specific behavior that Crash's rd command implements
for -sS[S]. Unfortunately, the addresses crash identifies and annotates
are a subset of what drgn is capable of annotating. It might be worth
including the entire set of drgn annotations with a dedicated option,
such as when specifying a third "-S" option. That is not implemneted
here.

Userspace memory reading is also supported, using the context task to
determine the address space. Since it's definitely non-trivial to
determine whether an address is user or kernel, we employ a trick: first
we try to read memory using prog.read, and if that fails, we try to read
via access_process_vm(). Based on that, we can "magically" generate the
correct drgn code. Usually the user knows whether they're trying to read
user/kernel memory, so it makes sense to pretend we magically knew it as
well.

The only crash options which are not supported are -m (for Xen host
machine addresses) and -f (for dumpfile offsets). Neither of these are
likely to be supported by drgn.

Signed-off-by: Stephen Brennan <stephen.s.brennan@oracle.com>