contrib/search_kernel_memory.py: handle excluded pages
search_kernel_memory.py currently doesn't work for many kernel core
dumps because they may have excluded pages which cause fault errors. We
can skip pages that fault, but to do that we also need to read in
page-sized chunks. This is noticeably slower for /proc/kcore, though.
So, start with a large chunk size, then fall back to page-by-page if we
hit a fault.

This is a stop-gap until we add a proper memory searching interface in
the next release.

Signed-off-by: Omar Sandoval <osandov@osandov.com>