drgn.helpers.linux.list: detect simple cycles in validate_list*
The validate_list family of helpers detect cycles in unchanging lists,
but if the list is being modified, they can miss a cycle and get stuck
in it. Handling these cases in general would require keeping a set of
all nodes we've seen, which is undesirable. We can handle the most
common case of a single-node cycle with a single extra check. This fixes
intermittent infinite loops in the
tests.linux_kernel.crash_commands.test_kmem.TestKmem.test_[fF] tests.

Signed-off-by: Omar Sandoval <osandov@osandov.com>