Temporarily hide commands
Commands and crash compatibility mode aren't quite ready to roll out
yet, but we have a bunch of fixes that need to be released. Leave all of
the command code in, but remove the drgn.commands._builtin plugin and
delete the documentation for now. Commands can still be accessed by
manually importing drgn.commands._builtin.

Signed-off-by: Omar Sandoval <osandov@osandov.com>