drgn.helpers.linux.timer: add {,hr}timer helpers
These are fairly minimal, just enough for the crash timer command. Since
the timer structures are kept internal, there's not much of a public
in-kernel API to copy for the helper API. So, I just documented code
snippets for how to get to the different timer bases. These could maybe
be adapted into helpers if desired.

Signed-off-by: Omar Sandoval <osandov@osandov.com>