drgn.helpers.common.format: support omitting aliases in decode_flags()
There are cases where the list of flags includes aliases of some flags,
but we only want to decode the "canonical" name (e.g., when decoding
sa_flags in the crash sig command). Add an "aliases" parameter to
decode_flags() to support this. The default remains the same (include
aliases).

Signed-off-by: Omar Sandoval <osandov@osandov.com>