vmtest.config: bump kernel.org crosstool compiler version
The binutils issue that caused empty segments (see commit 2a9053de8796
("libdrgn: elf_file: ignore empty program headers in
drgn_elf_file_address_range()")) appears to be fixed in the 12.5.0
toolchain. Upgrade it to the latest minor version. It'd be nice to
upgrade to a newer major version, but I don't want to risk a repeat of
commit c18a1a973e74 ("Revert "vmtest.config: bump kernel.org crosstool
compiler version"").

Signed-off-by: Omar Sandoval <osandov@osandov.com>