tests: crash: rd: fix annotation tests
The test_annotate_slab() and test_annotate_slab_verbose() cases rely on
the slab_caches variable to contain pointers to slab objects (those
allocated from the kmem_cache slab). This is normally the case, but it
can't be relied upon:

1. Some of the alternative configurations have "kmem_cache_boot" as the
head or the tail of this list. Since this is a symbol and not a slab
object, the test fails.
2. Kernel configurations with SLOB cannot detect which slab cache a
pointer belongs to, so they should output "unknown slab object". While I
would refer to crash's implementation to determine what it would output,
I cannot get crash to run on a relevant SLOB-enabled kernel. So, we'll
just define our own output - "unknown slab object".

Fixes: d6d9feb2 ("crash: port rd command")

Signed-off-by: Stephen Brennan <stephen.s.brennan@oracle.com>