vmtest: allow running specific tests in parallel
When doing development, it's nice to create a fast feedback loop, by
running just the specific tests related to the feature. This can be done
with a command like:

    python -m vmtest.vm --kernel-cmdline quiet \
                        --insert-test-kmod \
                        -k $KERNEL \
                        python3 -m pytest $FILE

However, this cannot be parallelized across multiple kernels the same
way that running the full test suite can. Incorporate this into the
vmtest runner, which supports parallelism, by adding the
"--pytest-kernel-args" option to specify pytest arguments, and also
"--skip-kdump" to avoid entering the kdump environment.

To make this a bit nicer, we also automatically add the "quiet" kernel
parameter when the test is in the foreground and we're specifying tests.
This feels like a reasonable heuristic to protect your terminal
scrollback. It also avoids polluting LLM context windows, if you're
attempting to use them.

Signed-off-by: Stephen Brennan <stephen.s.brennan@oracle.com>