tests: bpf: add instruction encoding functions
Our current BPF helper tests only use one program source, so it's not
too bad to hard-code the instructions for it (although it's a bit
awkward doing it for little- and big-endian). However, upcoming tests
will need at least one more instruction, so let's add constants and
functions for encoding BPF instructions mirroring the macros defined by
the kernel. I pulled in most of the constants but only added the needed
functions.

Signed-off-by: Omar Sandoval <osandov@osandov.com>