Add drgn.Program.source_location()/drgn_program_source_location()
For the crash sym command, we need a way to get the source code location
of a symbol. I have a hacky way to do that using
drgn.Program.stack_trace_from_pcs(), but recommending that in the --drgn
output feels wrong. Having a proper API to do this is a long-standing
feature request, anyways, so let's finally do it.

The implementation is a hack: drgn_stack_trace already provides all of
the necessary functionality, so this essentially re-exports it with a
different name and set of functions.

Closes #180.

Signed-off-by: Omar Sandoval <osandov@osandov.com>