libdrgn: ignore ELF symbols in relocatable files with missing section address
ELF symbols in relocatable files need to be relocated relative to their
corresponding section address. Currently, if a section address was not
reported (with drgn_module_set_section_address()), then we use whatever
the sh_addr value was in the ELF section header, which isn't meaningful.
As a result, we're returning symbols with arbitrary addresses even
though they're essentially undefined.

In practice, this mainly affects symbols coming from the .init* sections
of Linux kernel modules. These sections are unloaded after the module is
initialized, so the symbols don't really exist anymore.

We already ignore undefined symbols, i.e., symbols with an st_shndx of
SHN_UNDEF. So, let's ignore symbols from sections with a missing
address, too. To do so, we track which sections have a valid address in
a bitmap and check that bitmap.

Signed-off-by: Omar Sandoval <osandov@osandov.com>