Disallow module address range overlap
I originally wasn't sure whether to allow this or not. It doesn't make
sense for modules to have overlapping address ranges, but I also wanted
to be permissive in the case that a module was unloaded and then a new
module reused its address ranges. However, the module address tree
doesn't handle overlaps (it is keyed only on start address), so it was
always questionable. Even worse, inserting duplicate start addresses can
corrupt the tree:

1. Module A is created, set to start address X, and inserted in the
   tree.
2. Module B is created, set to start address Y, and inserted in the
   tree.
3. Module A is reset to start address Y.
   a. First, this deletes module A from the tree. Note that this doesn't
      reinitialize the tree node.
   b. Then, this attempts to insert module A at address Y. However,
      drgn_module_address_tree_insert() doesn't allow duplicates, so it
      doesn't get inserted. Instead, the node is left in whatever state
      it was in after it was deleted.
4. The tree is modified more.
5. Module A is reset to start address Z. This attempts to delete module
   A from the tree, but it's not actually in the tree, so deleting the
   node corrupts the tree.

So, let's be conservative and disallow overlap entirely. Hopefully this
doesn't break anyone, but it can be worked around by unsetting the
address range on the overlapping module.

Fixes #574.

Fixes: 4e83130008e9 ("Introduce module and debug info finder APIs")
Reported-by: Stephen Brennan <stephen.s.brennan@oracle.com>
Signed-off-by: Omar Sandoval <osandov@osandov.com>