libdrgn: python: make drgn.SourceLocation actually a NamedTuple
The SourceLocation class was intended to be fully backwards compatible
with the tuple previously returned by drgn.StackTrace.source(). However,
Stephen Brennan reported that it broke printf-style string formatting,
i.e., "%s:%d:%d" % frame.source(). This is because printf-style
formatting handles tuples specially, and although SourceLocation behaves
like a NamedTuple, it is not truly a tuple.

Fix it by replacing the custom class with an actual NamedTuple subclass.
This is tricky because the SourceLocation.name() method uses state
beyond the NamedTuple attributes. The least awful way to support this is
by creating a NamedTuple base class and then subclassing it in Python.
This results in instances with __dict__, allowing us to add the name
callback as an extra attribute.

Fixes: 53a0b3fa57ae ("Add class to represent stack frame source location")
Reported-by: Stephen Brennan <stephen.s.brennan@oracle.com>
Signed-off-by: Omar Sandoval <osandov@osandov.com>