python: implement drgn.Object.__reversed__()
I want to use reversed(obj) for some crash list command test cases, and
it's easy to support.

This also fixes a bug in _ObjectIterator.__length_hint__(): it should
return the number of items remaining in the iterator, not the total
number of items. Since it's just a hint, it doesn't matter much.

Signed-off-by: Omar Sandoval <osandov@osandov.com>