drgn.commands: split parsing and running of custom commands
Right now, custom commands that want to support redirections/pipes have
to do the actual redirection and piping themselves. This is annoying and
gets confusing with crash's default paging. Instead, split the parsing
and running, so that the parsing step can return a representation of
what redirections/pipes it wants for the command framework to actually
set up. custom_command() must now provide a parsing callback, and the
new raw_command() provides the old, unparsed behavior.

This also fixes #567 by ensuring that piping or redirecting stdout
disables the crash pager.

Signed-off-by: Omar Sandoval <osandov@osandov.com>