Add drgn-crash wrapper CLI
To hopefully make migration from drgn to crash even easier, add a
wrapper script that bypasses the drgn CLI completely and goes straight
into crash compatibility mode. The drgn CLI can still be entered with
the new "drgn" command.

Instead of arbitrary nesting of crash->drgn->crash->etc., if we're
already in a drgn REPL and attempt to enter drgn mode, we exit to the
outer REPL, and likewise for crash. The code for this is pretty ugly.

Closes #568.

Signed-off-by: Omar Sandoval <osandov@osandov.com>