drgn.cli: change history file location to $XDG_STATE_HOME/drgn/history
When we added a history file in commit 0a089e6ac9b9 ("cli: add history
and tab completion") in 2016, the XDG spec didn't have a standard place
to put history files, so we put it in ~/.drgn_history. However, the May
2021 revision of the XDG spec added XDG_STATE_HOME for that purpose. It
defaults to ~/.local/state.

Write new history files to $XDG_STATE_HOME/drgn/history, but continue
using the old path if it exists.

Closes #469.

Signed-off-by: Omar Sandoval <osandov@osandov.com>