tests: kmod: add test block device
For testing upcoming block layer helpers, we need to be able to inspect
requests while they are in flight. null_blk does not give us control
over when exactly requests are delayed and completed, so we need to
write our own block driver that gives us that control through sysfs and
an ioctl.

P.S. Writing a block driver that works on Linux 4.9-6.19 requires so
much conditional code.

Signed-off-by: Omar Sandoval <osandov@osandov.com>