crash: check validity of task struct arguments
Crash checks that a task argument given as a task_struct address is
valid and prints an error if not. Let's do the same, which we can do
relatively efficiently by checking that task->pid (or task_cpu(task) in
the case of PID 0) resolves back to the given task. _TaskSelector will
also be able to use this to avoid checking task_struct validity by
iterating over all tasks.

Signed-off-by: Omar Sandoval <osandov@osandov.com>