crash: support foreach ps
This one feels kind of redundant, but crash supports it. There are a few
flags that crash doesn't support with foreach but we do because it's
easier to allow them than it is to forbid them. Crash also allows
filtering further with -y or -G, so we hack that in, too.

Signed-off-by: Omar Sandoval <osandov@osandov.com>